requirements.txt
fastapi
uvicorn
python-dotenv
httpx
motor
streamlit